Update wording of 01-Getting-Started
Hi! I'm learning Javascript and have found these tutorials to be incredibly helpful. While the articles are very clear, some of the wording isn't natural to me as an English speaker. I was hoping I could contribute to this project by updating the wording of the articles to read more like natural English. If you like these edits, let me know and I'll make my way through the rest of the articles as I read them. Thanks for creating this.